pyproject.toml
[project]
name = "agent-scanner"
version = "0.1.0"
description = ""
authors = [
    {name = "Medhansh-741",email = "medhansh541@gmail.com"}
]
requires-python = ">=3.14"
dependencies = [
    "typer (>=0.24.1,<0.25.0)",
    "rich (>=14.3.3,<15.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "gitpython (>=3.1.46,<4.0.0)",
    "groq (>=1.0.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
